+++ src/components/manifold-marketplace-grid/manifold-marketplace-grid-happo.ts
@@ -1,8 +1,11 @@
+import skeletonProducts from '../../data/marketplace';
 import products from '../../spec/mock/products.json';
 import fromJSON from '../../spec/mock/fromJSON';
 
 export const skeleton = () => {
   const grid = document.createElement('manifold-marketplace-grid');
+  grid.services = fromJSON(skeletonProducts);
+  grid.skeleton = true;
 
   document.body.appendChild(grid);
 
@@ -12,6 +15,7 @@ export const skeleton = () => {
 export const allProducts = () => {
   const grid = document.createElement('manifold-marketplace-grid');
   grid.services = fromJSON(products);
+  grid.freeProducts = [];
 
   document.body.appendChild(grid);