+++ src/components/manifold-marketplace/manifold-marketplace.tsx
@@ -2,6 +2,7 @@ import { h, Component, Prop, State, Element, Watch } from '@stencil/core';
 
 import { Catalog } from '../../types/catalog';
 import Tunnel from '../../data/connection';
+import skeletonProducts from '../../data/marketplace';
 import { RestFetch } from '../../utils/restFetch';
 import logger from '../../utils/logger';
 
@@ -28,7 +29,7 @@ export class ManifoldMarketplace {
   @Prop() products?: string;
   /** Template format structure, with `:product` placeholder */
   @Prop() templateLinkFormat?: string;
-  @State() freeProducts: string[] = [];
+  @State() freeProducts?: string[];
   @State() parsedExcludes: string[] = [];
   @State() parsedFeatured: string[] = [];
   @State() parsedProducts: string[] = [];
@@ -40,7 +41,7 @@ export class ManifoldMarketplace {
 
   componentWillLoad() {
     this.parseProps();
-    this.fetchProducts();
+    this.fetchProducts(); // don’t wait on product fetch
   }
 
   fetchProducts = async () => {
@@ -63,6 +64,7 @@ export class ManifoldMarketplace {
   fetchFreeProducts = async () => {
     const freeProducts: string[] = [];
 
+    // Fetch all plans in parallel
     await Promise.all(
       this.services.map(
         ({ id }) =>
@@ -111,6 +113,8 @@ export class ManifoldMarketplace {
 
   @logger()
   render() {
+    const isLoading = !this.services.length || !this.freeProducts; // wait for free calls to finish
+
     return (
       <manifold-marketplace-grid
         excludes={this.parsedExcludes}
@@ -122,7 +126,8 @@ export class ManifoldMarketplace {
         preserveEvent={this.preserveEvent}
         productLinkFormat={this.productLinkFormat}
         products={this.parsedProducts}
-        services={this.services}
+        skeleton={isLoading}
+        services={isLoading ? skeletonProducts : this.services}
         templateLinkFormat={this.templateLinkFormat}
       />
     );