+++ src/components/manifold-resource-product/manifold-resource-product-happo.ts
@@ -1,0 +1,5 @@
+export const skeleton = () => {
+  const resourceProduct = document.createElement('manifold-resource-product');
+  document.body.appendChild(resourceProduct);
+  return resourceProduct.componentOnReady();
+};