+++ stories/manifold-marketplace.stories.js
@@ -7,21 +7,42 @@ storiesOf('Marketplace', module)
   .add(
     'default',
     () =>
-      '<manifold-marketplace product-link-format="/products/:product" template-link-format="/products/custom/:template" preserve-event></manifold-marketplace>'
+      `<manifold-marketplace
+        product-link-format="/products/:product"
+        preserve-events
+        template-link-format="/products/custom/:template"
+      ></manifold-marketplace>`
   )
-  .add('no templates', () => '<manifold-marketplace hide-templates></manifold-marketplace>')
+  .add(
+    'no templates',
+    () => `
+      <manifold-marketplace
+        hide-templates
+      ></manifold-marketplace>`
+  )
   .add(
     'featured',
-    () =>
-      '<manifold-marketplace featured="zerosix,jawsdb-maria,jawsdb-mysql,jawsdb-postgres"></manifold-marketplace>'
+    () => `
+      <manifold-marketplace
+        featured="zerosix,jawsdb-maria,jawsdb-mysql,jawsdb-postgres"
+      ></manifold-marketplace>`
   )
   .add(
     'compact',
-    () =>
-      '<manifold-marketplace hide-categories hide-search products="iron_cache,iron_mq,jawsdb-mysql,jawsdb-postgres,mailgun,logdna,iron_worker,zerosix"></manifold-marketplace>'
+    () => `
+      <manifold-marketplace
+        hide-categories
+        hide-search
+        products="iron_cache,iron_mq,jawsdb-mysql,jawsdb-postgres,mailgun,logdna,iron_worker,zerosix"
+      ></manifold-marketplace>`
   )
   .add(
     'curated',
-    () =>
-      '<manifold-marketplace hide-templates="" hide-search="" hide-categories="" products="logdna,scoutapp,timber-logging,cloudcube"></manifold-marketplace>'
+    () => `
+      <manifold-marketplace
+        hide-templates
+        hide-search
+        hide-categories
+        products="logdna,scoutapp,timber-logging,cloudcube"
+      ></manifold-marketplace>`
   );